commands (exept 'create') have been moved to a separate file
